acpi: add apple-specific hack to return "Darwin" for OSI
When running on machines with a hw_vendor of "Apple Inc." or "Apple
Computer, Inc.", only return 1 for an OSI check of "Darwin" and not
the other Windows variants.

Pretending to be Darwin seems to work better for getting the MBA7,1
to at least fully suspend (without it, _PTS(3) hangs).

However, since the DSDT on at least the MBA has a block that sets
OSYS like this:

        OSYS = 0x07DC
        If (CondRefOf (\_OSI, Local0))
        {
            If (_OSI ("Darwin"))
            {
                OSYS = 0x2710
            }

            If (\_OSI ("Linux"))
            {
                OSYS = 0x03E8
            }

            If (\_OSI ("Windows 2009"))
            {
                OSYS = 0x07D9
            }
            [...]

it's not enough to just add Darwin to aml_valid_osi, since it will
also return true for Windows variants.

This is what Linux has done for many years.