acpiec: busy sleep a little bit in acpiec_wait, before using tsleep
tsleep's minimum delay is still a bit long in the context of doing a
bunch of acpiec_{read,write} calls in a row, as acpisbs does

use delay() for a little while to try to read status quickly, before
falling back to tsleep.  long enough to hopefully fetch the status
quickly, but short enough that it won't cause noticeable delays to
userland, which is why tsleep is used here when !cold.

speeds up acpisbs's periodic smbus reads from ~3.2 seconds to 100ms