ubcmtp: present as WSMOUSE_TYPE_USB
this causes X not to open our unique /dev/wsmouse? device and
configure it with the synaptics driver, but to just keep talking to
it through the /dev/wsmouse mux device.

since most of the heavy lifting of two finger scrolling and other
stuff is done by wstpad now, we can get away with the xf86-input-ws
driver.

this fixes the problem of suspend and resume detaching and
reattaching wsmouse devices, but X not being able to see them