Prevent a NULL pointer deref in iwm(4) which I have seen during testing.
iwm_stop() sets the phy context pointer in the ic_bss node to NULL.
If iwm_stop() runs in parallel to the newstate task, the newstate task can
dereference this pointer in iwn_update_quotas() or iwm_binding_cmd().
So check the pointer for NULL before derefencing.

This is a quick and dirty workaround.
A proper fix for such task races is still pending.