acpiec: bring back non-bursting only on Google Chromebook Pixel
Old versions of the Google EC do not support burst mode and cause us
to spin forever waiting for the acknowledgment.

This was tried before but conflicted with Apple hardware and was
backed out.  Bring it back but only do it when the hw_vendor and
hw_prod match "GOOGLE Samus".