inteldrm: always discard BIOS fb size, not just if it's smaller
For some reason on the Macbook10,1, the BIOS reports a console size
of 2880x1800 (which is the actual resolution of the larger MacBook
Pro, maybe Apple just left that config in there) but the display
actually runs at 2304x1440.

When inteldrm takes over and tries to bring up a console in
2880x1800, the text is all screwed up.  However, once X11 loads and
uses the actual mode from probing the display (2304x1440), it draws
fine.  Switching back to the console again shows the corrupted text.

There is already code here to discard what the BIOS reports and use
the probed size, but only if the probed size is larger than what the
BIOS reported.  This change just makes it always do that when the
sizes do not match.