usbhidaction: add -u flag to specify a vendor and product
Using hotplug to monitor for USB devices, wait for a matching uhid
device to show up that matches the vendor, product, and report id.

Specifying a uhid path with -f can be difficult to know ahead of
time, especially for devices with USB keyboards and things that
detach and reattach at suspend/resume.