satopcase_send_msg: better handling of wait_reply
When we send a command and are waiting for a reply, we may get a
bunch of packets before our reply which are not errors.  This can
happen if the touchpad is being touched while booting, for example.

Instead of just consuming packets until we get a non-error, consume
until we receive a reply that matches the sent packet's message
type.