acpithinkpad: If CPU temp is low, turn fan off
The X1 Nano can usually cruise with no fan on during most normal
activity, but the EC has this temperature threshold set a tad too
low so it constantly toggles the fan between off and running at its
lowest speed.  Do this manually in the acpithinkpad driver but bump
the temperature threshold up a bit to stop it from turning on unless
needed.